+++ app/controllers/internal_controller.rb
@@ -13,7 +13,7 @@ class InternalController < ApplicationController
 
   def directory
     @order = scrub_order(Member, params[:order], 'members.last_name')
-    @members = Member.current.order(@order).includes(:site)
+    @members = Member.current.where(archived: false).order(@order).includes(:site)
   end
 
   def category