+++ db/migrate/20160509153602_add_archived_to_members.rb
@@ -1,0 +1,6 @@
+class AddArchivedToMembers < ActiveRecord::Migration
+  def change
+    add_column :members, :archived, :boolean, null: false, default: false
+    add_index :members, :archived
+  end
+end